Fix Reskinning of buttons that were skinned, but are in a disabled group
In WeakAuras we heavly reuse buttons even between groups. In the issue a
button that was previously in a enabled group is now moved to a disabled
group. In that case there's a skin on it (that is it has Regions), so
ReSkin should apply the default skin again.

Fixes WeakAuras Issue:
https://github.com/WeakAuras/WeakAuras2/issues/4168